Create agentslimbreg.py
Python version of the matlab limb regen simulation. Added some optimizations like vectorization and KDtrees. Going to continue optimizing as well.

You can set the profiling and video flags to true/false depending on if you want to generate video or profile.

NOTE: The code is messy, and there is dead code at the bottom, I will fix that soon.